requirementsUbuntu.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
h5py==3.7.0
idna==3.4
ipykernel==6.19.0
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter-client==7.4.8
jupyter-core==5.1.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
